@@ -6,9 +6,10 @@ import { RouterTestingModule } from '@angular/router/testing';
 import { ApiService } from 'src/app/services/api';
 import { KeycloakService } from 'src/app/services/keycloak.service';
 import { Router } from '@angular/router';
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
 import { ConfigService } from 'src/app/services/config.service';
 import { Utils } from 'src/app/shared/utils/utils';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 describe('LoginComponent', () => {
   let component: LoginComponent;
@@ -26,16 +27,18 @@ describe('LoginComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [LoginComponent],
-      imports: [FormsModule, RouterTestingModule, HttpClientTestingModule],
-      providers: [
+    declarations: [LoginComponent],
+    imports: [FormsModule, RouterTestingModule],
+    providers: [
         { provide: ApiService },
         { provide: KeycloakService, useValue: mockKeycloakService },
         { provide: Router, useValue: routerSpy },
         ConfigService,
-        Utils
-      ]
-    })
+        Utils,
+        provideHttpClient(withInterceptorsFromDi()),
+        provideHttpClientTesting()
+    ]
+})
       .compileComponents();
   }));
 
